services/usuarios/pyproject.toml
[project]
name = "usuarios"
version = "0.1.0"
description = "Microservicio de Usuarios"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "passlib[bcrypt]>=1.7.4,<2.0.0",
    "pydantic>=2.11.4",
    "pydantic-settings>=2.9.1",
    "pyjwt>=2.10.1",
    "python-jose[cryptography]>=3.4.0",
    "sqlmodel>=0.0.24",
    "typer>=0.15.3",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "ruff>=0.11.8",
]
[project.scripts]
usuarios_cli = "app.cli:app"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"
